file closing: takes into account the pending r/w req. before releasing
Because of the readahead it might be possible that there is some
read pendings. Incase of file cloe, Rozofs needs to wait for the
response otherwise the file descriptor might freed.
The case of the write was already addressed. Here we have a
protection against a potential write after close (not
really possible, but we need to protect Rozofs from that situation

Enhancement: when a read response is received, only the unread
part is filled in the fd's buffer. Rozofs provides fuse we
the pointer to requested section avoiding unnecessary data copy.