Description: ------------------ exportd and storcli run out of fd because there are a lot TCP connection established by storcli towards exportd that have not been closed.
Consequence:
---------------------
exportd and storcli are out of services (100% cpu usage)

Situation recovery
--------------------------
To recover from that situation we just have to kill  the storcli. The storcli
is automatically restarted and the exportd can service incoming
requests since fd have been released

Cause:
---------
Whenever there is a failure when the storcli attempts to get information about
the current exportd configuration (typically a timeout) the storcli re-opens
a TCP connection without checking the presence of the current one.
So after a while, the storcli runs out of fd, (same thing for exportd since
the storcli does not close the TCP connection.

The fix consists in testing the presence of the socket associated with the
TCP connexion. If there is one, the TCP connection is closed.
During the initialisation of the rpc client context, the reference of the socket,
CLIENT and lg_id MUST be cleared before starting the thread to avoid
closing a "fake" socket since the content of the context is unknown.