blockmode cache: take care of the beginning offset while writing data in cache:
The cache starts on a 8K boundary. When the data to put in the cache
does not start on a 8K boundary the data are not updated. This update
is rejected when the current block have not the same key as the one
that is already in cache. When they have the same key, the update can
take place.