Storcli ring: Bad control on FID. The side effect is a serialization of the request. The issue comes from a bid size comparison on the FID, only the 8 upper bytes were checked.
Lbg supervision:
When there is no response on a tcp connection, the system starts to
poll the connection. For this, a large buffer is allocated, however
when the transaction times-out, the system does not release the xmit
buffer. The system might run out of buffer. This can be observed in
the situation where the storio is break under gdb (not a real
situation). For sanity, the reference of the xmit buffer is saved
in the transaction context, thus upon times-out it is automatically released.